@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:ourlog/widgets/artwork_slider.dart';
 import 'package:ourlog/widgets/bulletin_board.dart';
 import 'package:ourlog/widgets/main_banner.dart';
-import '../services/artwork_service.dart';
+import '../services/picture_service.dart';
 import '../constants/theme.dart';
 import '../widgets/header.dart';
 import '../widgets/footer.dart';
